+++ src/sentineltray/__init__.py
@@ -1,5 +1,5 @@
 """SentinelTray package."""
 
-__version__ = "1.0.0b3"
-__version_label__ = "1.0.0-beta.3"
+__version__ = "1.0.0b4"
+__version_label__ = "1.0.0-beta.4"
 __release_date__ = "23-01-2026"